Refactor to add type::TypeId::REAL, scaffolding for type::TypeId::DECIMAL.

+++ src/binder/bind_node_visitor.cpp
@@ -838,7 +838,7 @@ void BindNodeVisitor::UnifyOrderByExpression(
         case type::TypeId::BIGINT:
           column_id = constant_value_expression->GetInteger().val_;
           break;
-        case type::TypeId::DECIMAL:
+        case type::TypeId::REAL:
           column_id = constant_value_expression->GetReal().val_;
           break;
         default:
